chore: align cross-namespace restore backport with main
