+++ config/crd/bases/dataprotection.kubeblocks.io_restores.yaml
@@ -85,7 +85,7 @@ spec:
                   namespace:
                     description: |-
                       Specifies the backup namespace.
-                      Cross-namespace VolumeSnapshot restores are not supported.
+                      Cross-namespace VolumeSnapshot restores are NOT supported.
                     type: string
                   sourceTargetName:
                     description: Specifies the source target for restoration, identified
