+++ deploy/helm/crds/operations.kubeblocks.io_opsrequests.yaml
@@ -621,7 +621,7 @@ spec:
                               description: |-
                                 Specifies the namespace of the Backup.
                                 If not specified, the namespace of the OpsRequest will be used.
-                                Cross-namespace VolumeSnapshot restores are not supported.
+                                Cross-namespace VolumeSnapshot restores are NOT supported.
                               type: string
                             restoreEnv:
                               description: |-
