+++ pkg/dataprotection/restore/utils_test.go
@@ -40,7 +40,7 @@ func TestValidateAndInitRestoreMGRCrossNamespaceBackup(t *testing.T) {
 	scheme := runtime.NewScheme()
 	assert.NoError(t, dpv1alpha1.AddToScheme(scheme))
 
-	fullActionSet := &dpv1alpha1.ActionSet{
+	actionSet := &dpv1alpha1.ActionSet{
 		ObjectMeta: metav1.ObjectMeta{Name: "full-action"},
 		Spec: dpv1alpha1.ActionSetSpec{
 			BackupType: dpv1alpha1.BackupTypeFull,
@@ -51,7 +51,7 @@ func TestValidateAndInitRestoreMGRCrossNamespaceBackup(t *testing.T) {
 		ObjectMeta: metav1.ObjectMeta{Name: "backup", Namespace: "source"},
 		Status: dpv1alpha1.BackupStatus{
 			Phase:        dpv1alpha1.BackupPhaseCompleted,
-			BackupMethod: &dpv1alpha1.BackupMethod{Name: "full", ActionSetName: fullActionSet.Name},
+			BackupMethod: &dpv1alpha1.BackupMethod{Name: "full", ActionSetName: actionSet.Name},
 		},
 	}
 	restoreObj := &dpv1alpha1.Restore{
@@ -64,9 +64,8 @@ func TestValidateAndInitRestoreMGRCrossNamespaceBackup(t *testing.T) {
 		Ctx: context.Background(),
 		Req: ctrl.Request{NamespacedName: client.ObjectKeyFromObject(restoreObj)},
 	}
-
 	t.Run("allows the Backup reference by default", func(t *testing.T) {
-		cli := fake.NewClientBuilder().WithScheme(scheme).WithObjects(fullActionSet, backup).Build()
+		cli := fake.NewClientBuilder().WithScheme(scheme).WithObjects(actionSet, backup).Build()
 		mgr := &RestoreManager{Restore: restoreObj.DeepCopy()}
 		assert.NoError(t, ValidateAndInitRestoreMGR(reqCtx, cli, mgr))
 		assert.Len(t, mgr.PrepareDataBackupSets, 1)
@@ -74,9 +73,9 @@ func TestValidateAndInitRestoreMGRCrossNamespaceBackup(t *testing.T) {
 
 	t.Run("rejects a cross-namespace VolumeSnapshot restore", func(t *testing.T) {
 		snapshotBackup := backup.DeepCopy()
-		snapshotVolumes := true
-		snapshotBackup.Status.BackupMethod.SnapshotVolumes = &snapshotVolumes
-		cli := fake.NewClientBuilder().WithScheme(scheme).WithObjects(fullActionSet, snapshotBackup).Build()
+		snapshotBackup.Status.BackupMethod.SnapshotVolumes = new(bool)
+		*snapshotBackup.Status.BackupMethod.SnapshotVolumes = true
+		cli := fake.NewClientBuilder().WithScheme(scheme).WithObjects(actionSet, snapshotBackup).Build()
 		err := ValidateAndInitRestoreMGR(reqCtx, cli, &RestoreManager{Restore: restoreObj.DeepCopy()})
 		assert.ErrorContains(t, err, "VolumeSnapshot")
 		assert.True(t, intctrlutil.IsTargetError(err, intctrlutil.ErrorTypeFatal))
@@ -115,7 +114,7 @@ func TestValidateAndInitRestoreMGRCrossNamespaceBackup(t *testing.T) {
 				},
 				BackupMethod: &dpv1alpha1.BackupMethod{
 					Name:            "snapshot",
-					ActionSetName:   fullActionSet.Name,
+					ActionSetName:   actionSet.Name,
 					SnapshotVolumes: &snapshotVolumes,
 				},
 			},
@@ -138,7 +137,7 @@ func TestValidateAndInitRestoreMGRCrossNamespaceBackup(t *testing.T) {
 		pitrRestore := restoreObj.DeepCopy()
 		pitrRestore.Spec.Backup = dpv1alpha1.BackupRef{Name: continuousBackup.Name, Namespace: continuousBackup.Namespace}
 		pitrRestore.Spec.RestoreTime = now.Add(-30 * time.Minute).Format(time.RFC3339)
-		cli := fake.NewClientBuilder().WithScheme(scheme).WithObjects(fullActionSet, continuousActionSet, baseBackup, continuousBackup).Build()
+		cli := fake.NewClientBuilder().WithScheme(scheme).WithObjects(actionSet, continuousActionSet, baseBackup, continuousBackup).Build()
 
 		err := ValidateAndInitRestoreMGR(reqCtx, cli, &RestoreManager{Restore: pitrRestore})
 		assert.ErrorContains(t, err, "VolumeSnapshot")
